Add date validation and refactor books and labels
