+++ app.rb
@@ -1,21 +1,24 @@
-require_relative './books/book_label_methods'
+require_relative './books/book_methods'
 require_relative './books/book'
 require_relative './games/game_methods'
 require_relative './games/game'
 require_relative './authors/author_methods'
 require_relative './authors/author'
+require_relative './labels/label'
+require_relative './labels/label_methods'
 require 'json'
 
 class App
   def initialize
-    @books = []
-    @labels = []
+    @books = load_all_books
+    @labels = load_all_labels
     @genres = []
     @authors = []
     @games = []
   end
 
-  include BookAndLabelMethods
+  include BookMethods
+  include LabelMethods
   include GameMethods
   include AuthorMethods
