+++ books/book.rb
@@ -5,14 +5,9 @@ class Book < Item
 
   def initialize(title, publisher, cover_state, publish_date)
     super(publish_date)
-    @title = capitalize_word(title)
-    @publisher = capitalize_word(publisher)
+    @title = title
+    @publisher = publisher
     @cover_state = cover_state
-  end
-
-  def capitalize_word(val)
-    val = val.split.each(&:capitalize!)
-    val.join(' ')
   end
 
   def can_be_archived?
