+++ books/books.json
@@ -1,0 +1,1 @@
+[{"title":"Awesome Book","publish_date":"2022-09-09","publisher":"Virag","cover_state":"good"}]
