+++ games/games.json
@@ -1,0 +1,4 @@
+[{"multiplayer":"true","last_played_at":"2022-09-09","publish_date":"2022-07-06","genre":{
+    "id": "3642354b9a",
+    "name":"fun"
+     },"label":"#<Label:0x00007fdefebde9c8>","author":"#<Author:0x00007fdefebde928>"}]
