--- labels.json
@@ -1,1 +1,0 @@
-[{"title":null,"color":"yellow"},{"title":"GIFT","color":"grey"},{"title":"new","color":"white"},{"title":"old","color":"brown"},{"title":"gift","color":"pink"},{"title":"old","color":"black"}]
