+++ labels/label.rb
