+++ labels/labels.json
@@ -1,0 +1,1 @@
+[{"title":"new","color":"pink"}]
